Move log_utils to logging
