+++ scripts/tests/chiptest/logging.py
